+++ src/pages/page-names.constants.ts
@@ -12,7 +12,7 @@ export const PASS_FINALISATION_PAGE = 'PassFinalisationPage';
 export const REKEY_SEARCH_PAGE = 'RekeySearchPage';
 export const TEST_REPORT_PAGE = 'TestReportPage';
 export const TEST_RESULTS_SEARCH_PAGE = 'TestResultsSearchPage';
-export const WAITING_ROOM_PAGE = 'WaitingRoomPage';
+export const WAITING_ROOM_PAGE = 'CatBWaitingRoomPage';
 export const WAITING_ROOM_TO_CAR_PAGE = 'WaitingRoomToCarPage';
 export const VIEW_TEST_RESULT_PAGE = 'ViewTestResultPage';
 export const JOURNAL_FORCE_CHECK_MODAL = 'JournalForceCheckModal';
