+++ src/pages/waiting-room/cat-b/cat-b-waiting-room.html
