+++ src/pages/waiting-room/cat-b/cat-b-waiting-room.scss
@@ -1,4 +1,4 @@
-page-waiting-room {
+page-cat-b-waiting-room {
 
   ion-content > div {
     background-color: map-get($colors, "mes-white");
@@ -34,7 +34,7 @@ page-waiting-room {
     .signature-validation-row {
       height: 300px;
     }
-    
+
     conducted-language > ion-row {
       margin-bottom: -36px;
     }
